ipc4: add ipc4 message support
There are two types of ipc4 message: global message and module message.
Ipc message is first decoded in handler.c and processed with the help
of helper.c to deal with pipeline message and module message.
Pipeline message and Module messge are converted to sof pipeline and
component message.

Currently only part of ipc4 message are processed and others will be done
in future.

Signed-off-by: Rander Wang <rander.wang@intel.com>